Move pol code around so it's cleaner.
